wso2project/requirements.txt
Django==4.2.4
django-import-export==3.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.31.0